Move app and form into components folder
