+++ src/components/Form/Form.js
@@ -1,0 +1,23 @@
+import React, { Component } from 'react';
+
+class Form extends Component {
+  constructor() {
+    super();
+    this.state = {};
+  }
+
+  render() {
+    return (
+      <form>
+        <input
+          type="text"
+
+        />
+      </form>
+    )
+  }
+
+
+}
+
+export default Form
